refactor: disable automated video cleanup job and integrate deletion logic into content cleanup workflow

+++ src/services/script.py
@@ -21,41 +21,41 @@ class ScriptService(BaseService):
         self.content_service = ContentService()
         self.short_video_service = ShortVideoService()
 
-    def clean_uploaded_video(self):
-        logger.info(f"Starting to clean uploaded videos")
-        try:
-            logger.info(f"Step 1/4: Getting videos")
-            videos = self.short_video_service.get_short_video_by_status(
-                self.db, short_video_status.ShortVideoStatus.PUBLISHED
-            )
-            logger.info(f"Step 2/4: Got videos: {len(videos)} videos")
-            logger.info(f"Step 3/4: Cleaning videos")
-            for video in videos:
-                logger.info(f"Cleaning video: {video.id}")
-                if os.path.exists(video.output_path):
-                    logger.info(f"Deleting video: {video.output_path}")
-                    os.remove(video.output_path)
-                    logger.info(f"Deleted video: {video.output_path}")
-                if os.path.exists(video.content.audio_path):
-                    logger.info(f"Deleting video: {video.content.audio_path}")
-                    os.remove(video.content.audio_path)
-                    logger.info(f"Deleted video: {video.content.audio_path}")
-                if os.path.exists(video.background_video_url):
-                    logger.info(f"Deleting video: {video.background_video_url}")
-                    os.remove(video.background_video_url)
-                    logger.info(f"Deleted video: {video.background_video_url}")
-                logger.info(f"Deleting video from db: {video.id}")
-                content_crud.delete_content(self.db, video.content)
-                logger.info(f"Deleted video from db: {video.id}")
-                short_video_crud.delete_short_video(self.db, video)
-                logger.info(f"Deleted video from db: {video.id}")
-            logger.info(f"Step 4/4: Deleted videos")
-        except Exception as e:
-            if isinstance(e, HTTPException):
-                logger.error(f"Error 1: {e.status_code} - {e.detail}")
-            else:
-                logger.error(f"Error 1: {str(e)}")
-            return
+    # def clean_uploaded_video(self):
+    #     logger.info(f"Starting to clean uploaded videos")
+    #     try:
+    #         logger.info(f"Step 1/4: Getting videos")
+    #         videos = self.short_video_service.get_short_video_by_status(
+    #             self.db, short_video_status.ShortVideoStatus.PUBLISHED
+    #         )
+    #         logger.info(f"Step 2/4: Got videos: {len(videos)} videos")
+    #         logger.info(f"Step 3/4: Cleaning videos")
+    #         for video in videos:
+    #             logger.info(f"Cleaning video: {video.id}")
+    #             if os.path.exists(video.output_path):
+    #                 logger.info(f"Deleting video: {video.output_path}")
+    #                 os.remove(video.output_path)
+    #                 logger.info(f"Deleted video: {video.output_path}")
+    #             if os.path.exists(video.content.audio_path):
+    #                 logger.info(f"Deleting video: {video.content.audio_path}")
+    #                 os.remove(video.content.audio_path)
+    #                 logger.info(f"Deleted video: {video.content.audio_path}")
+    #             if os.path.exists(video.background_video_url):
+    #                 logger.info(f"Deleting video: {video.background_video_url}")
+    #                 os.remove(video.background_video_url)
+    #                 logger.info(f"Deleted video: {video.background_video_url}")
+    #             logger.info(f"Deleting video from db: {video.id}")
+    #             content_crud.delete_content(self.db, video.content)
+    #             logger.info(f"Deleted video from db: {video.id}")
+    #             short_video_crud.delete_short_video(self.db, video)
+    #             logger.info(f"Deleted video from db: {video.id}")
+    #         logger.info(f"Step 4/4: Deleted videos")
+    #     except Exception as e:
+    #         if isinstance(e, HTTPException):
+    #             logger.error(f"Error 1: {e.status_code} - {e.detail}")
+    #         else:
+    #             logger.error(f"Error 1: {str(e)}")
+    #         return
 
     def clean_last_7_days_log_file(self):
         logger.info("Starting log cleanup")
@@ -103,12 +103,12 @@ class ScriptService(BaseService):
         logger.info(f"Starting to create content")
         logger.info(f"Step 1/3: Getting topic")
         topic = self.topic_service.get_random_unused_topic()
-        logger.info(f"Step 2/3: Got topic: {topic}")
 
         try:
             if not topic:
-                logger.info(f"No topic found to process")
+                logger.error(f"No topic found to process")
                 return
+            logger.info(f"Step 2/3: Got topic: {topic.id}")
             logger.info(f"Step 3/3: Creating content with topic")
             self.content_service.create_content(topic.id)
             logger.info(f"Step 3/3: Created content with topic: {topic.id}")
@@ -250,11 +250,33 @@ class ScriptService(BaseService):
                     or content.status == content_status.ContentStatus.ERROR
                 ):
                     logger.info(f"Cleaning content {content.id}")
-                    if content.audio_path is not None:
+                    if content.audio_path is not None and os.path.exists(content.audio_path):
                         os.remove(content.audio_path)
                         logger.info(f"Cleaned content {content.id} audio path")
                     content_crud.delete_content(self.db, content)
                     logger.info(f"Cleaned content {content.id}")
+                elif (len(content.short_video) > 0 and content.status == content_status.ContentStatus.VIDEO_PUBLISHED):
+                    logger.info(f"Content {content.id} is published")
+                    if len(content.short_video) > 0:
+                        for short_video in content.short_video:
+                            if short_video.status == short_video_status.ShortVideoStatus.PUBLISHED:
+                                logger.info(f"Cleaning video: {short_video.id}")
+                                if short_video.output_path and os.path.exists(short_video.output_path):
+                                    logger.info(f"Deleting video: {short_video.output_path}")
+                                    os.remove(short_video.output_path)
+                                    logger.info(f"Deleted video: {short_video.output_path}")
+                                if short_video.content.audio_path and os.path.exists(short_video.content.audio_path):
+                                    logger.info(f"Deleting video: {short_video.content.audio_path}")
+                                    os.remove(short_video.content.audio_path)
+                                    logger.info(f"Deleted video: {short_video.content.audio_path}")
+                                if short_video.background_video_url and os.path.exists(short_video.background_video_url):
+                                    logger.info(f"Deleting video: {short_video.background_video_url}")
+                                    os.remove(short_video.background_video_url)
+                                    logger.info(f"Deleted video: {short_video.background_video_url}")
+                                logger.info(f"Deleting video from db: {short_video.id}")
+                                short_video_crud.delete_short_video(self.db, short_video)
+                                logger.info(f"Deleted video from db: {short_video.id}")
+                    content_crud.delete_content(self.db, short_video.content)
             logger.info(f"Step 4/4: Cleaned last 7 days contents")
         except Exception as e:
             if isinstance(e, HTTPException):
